chore(jans-lock): Massively improved project structure
