+++ jans-lock/cedarling/README.md
@@ -1,0 +1,5 @@
+## cedarling
+
+The `cedarling` is an embeddable Webassembly Component that runs a local Cedar Engine, enabling fine grained and responsive Policy Management on the Web. The `cedarling` allows for dynamic updates to it's internal Policy Store via Server Sent events, enabling sub-second Access Management.
+
+### How it works
